Update 1a, 1b on change
Merge node updates
Use single state entry (state.data)
ODP cell listen on odp changes
